+++ selfdrive/ui/qt/onroad.h
@@ -155,6 +155,7 @@ signals:
   void mapPanelRequested();
 
 private:
+  void createMapWidget();
   void paintEvent(QPaintEvent *event);
   void mousePressEvent(QMouseEvent* e) override;
   OnroadAlerts *alerts;