+++ RELEASES.md
@@ -1,4 +1,4 @@
-Version 0.9.7 (2024-XX-XX)
+Version 0.9.7 (2024-05-XX)
 ========================
 * AleSato stuff
   * Change follows distances with long press on distance button (2024-03-13)
@@ -7,6 +7,8 @@ Version 0.9.7 (2024-XX-XX)
 * New driving model
 * Adjust driving personality with the follow distance button
 * Support for hybrid variants of supported Ford models
+* Added toggle to enable driver monitoring even when openpilot is not engaged
+* Fingerprinting without the OBD-II port on all cars
 
 Version 0.9.6 (2024-02-27)
 ========================